feat: demo /features concept with two-things
