--- src/drawings/two-things/index.js
@@ -1,1 +1,0 @@
-export { TwoThings } from './TwoThings'
