+++ src/features/two-things/index.js
@@ -1,0 +1,1 @@
+export { default as TwoThings } from './TwoThings'
